(PC-10067) Add tests native for VenueTile
