+++ src/features/home/atoms/tests/OfferTile.web.test.tsx
@@ -39,7 +39,7 @@ describe('OfferTile component', () => {
   it.skip('should navigate to the offer when clicking on the image [WEB INTEGRATION REQUIRED]', async () => {
     // eslint-disable-next-line local-rules/no-react-query-provider-hoc
     const { getByTestId } = render(reactQueryProviderHOC(<OfferTile {...props} />))
-    fireEvent.click(getByTestId('offerTileImage'))
+    fireEvent.click(getByTestId('tileImage'))
     expect(navigate).toHaveBeenCalledWith('Offer', {
       id: offerId,
       from: 'home',
@@ -51,7 +51,7 @@ describe('OfferTile component', () => {
   it.skip('Analytics - should log ConsultOffer that user opened the offer [WEB INTEGRATION REQUIRED]', async () => {
     // eslint-disable-next-line local-rules/no-react-query-provider-hoc
     const { getByTestId } = render(reactQueryProviderHOC(<OfferTile {...props} />))
-    fireEvent.click(getByTestId('offerTileImage'))
+    fireEvent.click(getByTestId('tileImage'))
     expect(analytics.logConsultOffer).toHaveBeenCalledWith({
       offerId,
       from: 'home',
@@ -63,7 +63,7 @@ describe('OfferTile component', () => {
   it.skip('should prepopulate react-query cache when clicking on offer [WEB INTEGRATION REQUIRED]', async () => {
     // eslint-disable-next-line local-rules/no-react-query-provider-hoc
     const { getByTestId } = render(reactQueryProviderHOC(<OfferTile {...props} />))
-    fireEvent.click(getByTestId('offerTileImage'))
+    fireEvent.click(getByTestId('tileImage'))
 
     const queryHash = JSON.stringify(['offer', offerId])
     const query = queryCache.get(queryHash)
