+++ src/features/home/components/VenuesModule.tsx
@@ -23,7 +23,7 @@ export const VenuesModule = (props: VenuesModuleProps) => {
   const renderItem: ListRenderItem<VenueHit> = useCallback(({ item }) => {
     return (
       <VenueTile
-        venueId={+item.id}
+        venueId={Number(item.id)}
         name={item.name}
         venueType={item.venueType}
         distance={formatDistance(item.position, userPosition)}
