+++ src/features/venue/atoms/__tests__/VenueOfferTile.native.test.tsx
@@ -38,7 +38,7 @@ describe('VenueOfferTile component', () => {
   it('should navigate to the offer when clicking on the image', async () => {
     // eslint-disable-next-line local-rules/no-react-query-provider-hoc
     const { getByTestId } = render(reactQueryProviderHOC(<VenueOfferTile {...props} />))
-    fireEvent.press(getByTestId('offerTileImage'))
+    fireEvent.press(getByTestId('tileImage'))
     expect(navigate).toHaveBeenCalledWith('Offer', {
       id: offerId,
       from: 'venue',
@@ -48,7 +48,7 @@ describe('VenueOfferTile component', () => {
   it('Analytics - should log ConsultOffer that user opened the offer', async () => {
     // eslint-disable-next-line local-rules/no-react-query-provider-hoc
     const { getByTestId } = render(reactQueryProviderHOC(<VenueOfferTile {...props} />))
-    fireEvent.press(getByTestId('offerTileImage'))
+    fireEvent.press(getByTestId('tileImage'))
     expect(analytics.logConsultOffer).toHaveBeenNthCalledWith(1, {
       offerId,
       from: 'venue',
@@ -59,7 +59,7 @@ describe('VenueOfferTile component', () => {
   it('should prepopulate react-query cache when clicking on offer', async () => {
     // eslint-disable-next-line local-rules/no-react-query-provider-hoc
     const { getByTestId } = render(reactQueryProviderHOC(<VenueOfferTile {...props} />))
-    fireEvent.press(getByTestId('offerTileImage'))
+    fireEvent.press(getByTestId('tileImage'))
 
     const queryHash = JSON.stringify(['offer', offerId])
     const query = queryCache.get(queryHash)
